Rudimentary support for finding first equation for a specified level.
For some reason it doesn't work if you specify a goal level twice in a row. You need to find the first 10 equations in each level fist before you can specify a goal level again. Dunno why. Probably something to do with the collection of permutations not being cleared or something idk.